Update Delete Contributor

+++ src/app/menus/manageusers/viewcontributor/viewcontributor.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit, Inject } from '@angular/core';
 import { ContributorService } from '../../../services/contributor.service';
-import { MatDialog ,MatDialogRef, MAT_DIALOG_DATA} from '@angular/material';
+import { MatDialog } from '@angular/material';
+import { NotificationService } from 'src/app/services/notification.service';
+import { DialogService } from 'src/app/services/dialog.service';
 
 @Component({
   selector: 'app-viewcontributor',
@@ -10,7 +12,11 @@ import { MatDialog ,MatDialogRef, MAT_DIALOG_DATA} from '@angular/material';
 export class ViewcontributorComponent implements OnInit {
 
   contributors: Array<any>;
-  constructor(private contributorService : ContributorService,public dialog: MatDialog){}
+  constructor(
+    private contributorService : ContributorService ,
+    private dialog: MatDialog,
+    private notificationService: NotificationService,
+    private dialogService: DialogService) { }
 
   ngOnInit() {
     this.contributorService.getAllContributors().subscribe(data=>{
@@ -18,8 +24,14 @@ export class ViewcontributorComponent implements OnInit {
       });
   }
 
-  remove(id) {
-      this.contributorService.deleteContributor(id).subscribe(result => {
-      }, error => console.error(error));   
+  remove(id:string){
+    this.dialogService.openConfirmDialog('Are you sure to delete this record ?')
+    .afterClosed().subscribe(res =>{
+      if(res){
+        this.contributorService.deleteContributor(id).subscribe(result => {
+        }, error => console.error(error));  
+        this.notificationService.warn('! Deleted successfully');
+      }
+    });
   }
 }
